backend/requirements.txt
absl-py
anyio
astunparse
cachetools
certifi
charset-normalizer
click
colorama
fastapi
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11
h5py
idna
keras
Keras-Preprocessing
libclang
Markdown
MarkupSafe
numpy
oauthlib
opt-einsum
packaging
Pillow
protobuf
pyasn1
pyasn1-modules
pydantic
pyparsing
python-multipart
requests
requests-oauthlib
rsa
six
sniffio
starlette
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
typing_extensions
unicorn
urllib3
uvicorn
Werkzeug
wrapt
